[chore] renamed domain to model for better understanding
