+++ backend/src/main/java/de/kruemelnerd/todo/backend/repository/TodoRepository.java
@@ -1,6 +1,6 @@
 package de.kruemelnerd.todo.backend.repository;
 
-import de.kruemelnerd.todo.backend.domain.Todo;
+import de.kruemelnerd.todo.backend.model.Todo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
