+++ backend/src/test/java/de/kruemelnerd/todo/backend/controller/TodoControllerITTest.java
@@ -1,6 +1,6 @@
 package de.kruemelnerd.todo.backend.controller;
 
-import de.kruemelnerd.todo.backend.domain.Todo;
+import de.kruemelnerd.todo.backend.model.Todo;
 import de.kruemelnerd.todo.backend.repository.TodoService;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
